+++ src/GWind/StorageBucket/Control.hs
@@ -1,22 +1,34 @@
--- ⚠
+--
 module GWind.StorageBucket.Control (reviseStorageBucketsM, createStorageBucketM, deleteStorageBucketM)
 where
 
     import GWind.Project.Domain (ProjectSnake)
+    import GWind.Project.Mutator (projectId)
+    import GWind.StorageBucket.Mutator (storageBucketId)
+    import GWind.StorageBucket.Service (revisionStorageBucketsM, creationCodeM, deletionCodeM)
+    import GWind.Common.Presenter (presentingM)
 
 --
     reviseStorageBucketsM :: ProjectSnake -> IO ()
-    reviseStorageBucketsM _ =
-        undefined
+    reviseStorageBucketsM projectSnake =
+        let pid = projectId projectSnake
+        in revisionStorageBucketsM pid
+            >>= presentingM
 
 --
     createStorageBucketM :: ProjectSnake -> IO ()
-    createStorageBucketM _ =
-        undefined
+    createStorageBucketM projectSnake =
+        let
+            pid = projectId projectSnake
+            sid = storageBucketId projectSnake
+        in creationCodeM pid sid
+            >>= presentingM
 
 --
     deleteStorageBucketM :: ProjectSnake -> IO ()
-    deleteStorageBucketM _ =
-        undefined
+    deleteStorageBucketM projectSnake =
+        let sid = storageBucketId projectSnake
+        in deletionCodeM sid
+            >>= presentingM
 
 --