+++ src/GWind/StorageBucket/Domain.hs
@@ -1,0 +1,9 @@
+-- ⚠
+module GWind.StorageBucket.Domain (StorageBucketId)
+where
+
+    import GWind.Common.Domain (Kebab)
+
+    type StorageBucketId = Kebab
+
+--