+++ src/GWind/StorageBucket/Mutator.hs
@@ -1,0 +1,14 @@
+--
+module GWind.StorageBucket.Mutator (storageBucketId)
+where
+
+    import GWind.Common.Mutator (kebab)
+    import GWind.Project.Domain (ProjectSnake)
+    import GWind.StorageBucket.Domain (StorageBucketId)
+
+--
+    storageBucketId :: ProjectSnake -> StorageBucketId
+    storageBucketId projectSnake =
+        kebab projectSnake ++ "-terraform"
+
+--